EICNET-813: Move logic from EntityOperations class to Pathauto class; Re-factor some code in Pathauto class to meet grump requirements.
