+++ lib/modules/eic_groups/src/Hooks/EntityOperations.php
@@ -7,9 +7,7 @@ use Drupal\Core\Entity\Display\EntityViewDisplayInterface;
 use Drupal\Core\Entity\EntityInterface;
 use Drupal\Core\Entity\EntityTypeManagerInterface;
 use Drupal\Core\Link;
-use Drupal\Core\Queue\QueueFactory;
 use Drupal\Core\Routing\RouteMatchInterface;
-use Drupal\Core\State\StateInterface;
 use Drupal\Core\StringTranslation\StringTranslationTrait;
 use Drupal\Core\Url;
 use Drupal\eic_groups\EICGroupsHelperInterface;
@@ -57,20 +55,6 @@ class EntityOperations implements ContainerInjectionInterface {
   protected $pathautoGenerator;
 
   /**
-   * The queue factory service.
-   *
-   * @var \Drupal\Core\Queue\QueueFactory
-   */
-  protected $queueFactory;
-
-  /**
-   * The state service.
-   *
-   * @var \Drupal\Core\State\StateInterface
-   */
-  protected $state;
-
-  /**
    * Constructs a new EntityOperations object.
    *
    * @param \Drupal\Core\Entity\EntityTypeManagerInterface $entity_type_manager
@@ -81,18 +65,12 @@ class EntityOperations implements ContainerInjectionInterface {
    *   The EIC Groups helper service.
    * @param \Drupal\pathauto\PathautoGeneratorInterface $pathauto_generator
    *   The pathauto generator.
-   * @param \Drupal\Core\Queue\QueueFactory $queue_factory
-   *   The queue factory service.
-   * @param \Drupal\Core\State\StateInterface $state
-   *   The state service.
    */
-  public function __construct(EntityTypeManagerInterface $entity_type_manager, RouteMatchInterface $route_match, EICGroupsHelperInterface $eic_groups_helper, PathautoGeneratorInterface $pathauto_generator, QueueFactory $queue_factory, StateInterface $state) {
+  public function __construct(EntityTypeManagerInterface $entity_type_manager, RouteMatchInterface $route_match, EICGroupsHelperInterface $eic_groups_helper, PathautoGeneratorInterface $pathauto_generator) {
     $this->entityTypeManager = $entity_type_manager;
     $this->routeMatch = $route_match;
     $this->eicGroupsHelper = $eic_groups_helper;
     $this->pathautoGenerator = $pathauto_generator;
-    $this->queueFactory = $queue_factory;
-    $this->state = $state;
   }
 
   /**
@@ -103,9 +81,7 @@ class EntityOperations implements ContainerInjectionInterface {
       $container->get('entity_type.manager'),
       $container->get('current_route_match'),
       $container->get('eic_groups.helper'),
-      $container->get('pathauto.generator'),
-      $container->get('queue'),
-      $container->get('state')
+      $container->get('pathauto.generator')
     );
   }
 
@@ -123,12 +99,6 @@ class EntityOperations implements ContainerInjectionInterface {
     // Publish group wiki when group is published.
     if (!$entity->original->isPublished() && $entity->isPublished()) {
       $this->publishGroupWiki($entity);
-    }
-
-    // If group alias has changed we add the group id into a queue so that all
-    // group content url aliases can be updated at a later stage with cron.
-    if ($entity->get('path')->alias !== $this->pathautoGenerator->createEntityAlias($entity, 'return')) {
-      $this->createGroupUrlAliasUpdateQueueItem($entity);
     }
   }
 
@@ -253,22 +223,6 @@ class EntityOperations implements ContainerInjectionInterface {
         $node_book->setPublished();
         $node_book->save();
       }
-    }
-  }
-
-  /**
-   * Creates an item in the queue CronOperations::GROUP_URL_ALIAS_UPDATE_QUEUE.
-   *
-   * @param \Drupal\group\Entity\GroupInterface $entity
-   *   The group entity object.
-   */
-  private function createGroupUrlAliasUpdateQueueItem(GroupInterface $entity) {
-    if (is_null($this->state->get(CronOperations::GROUP_URL_ALIAS_UPDATE_STATE_CACHE . $entity->id()))) {
-      $queue = $this->queueFactory->get(CronOperations::GROUP_URL_ALIAS_UPDATE_QUEUE);
-      $queue->createItem([
-        'gid' => $entity->id(),
-      ]);
-      $this->state->set(CronOperations::GROUP_URL_ALIAS_UPDATE_STATE_CACHE . $entity->id(), TRUE);
     }
   }
